useNativeLazyload function merged with MiniLazyload
